src/model_deployed/requirements.txt
asgi-correlation-id==4.3.4
fastapi==0.115.8
numpy==2.2.3
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.11.0.86
pre_commit==4.1.0
pydantic==2.10.6
pydantic_settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
scikit-image==0.25.1
structlog==25.1.0
uvicorn==0.34.0
